About Lunch Ticker Form
This is the source code for an application I am currently working on, for the kitchen of the NFP organization I am a part of. The idea is to keep track of lunch participants on a given day, to help reduce food waste when lunch is prepared. It is still a work in progress and I hope to extend the program over periods of time. I still have plans to add a statistical component to it, as well as an system ID tracker to prevent spamming of the staff button.